refactor: 면접 질문 도메인 위치 변경 #101
