--- src/main/java/KUSITMS/WITHUS/domain/application/interviewQuestion/repository/InterviewQuestionRepository.java
@@ -1,7 +1,0 @@
-package KUSITMS.WITHUS.domain.application.interviewQuestion.repository;
-
-import KUSITMS.WITHUS.domain.application.interviewQuestion.entity.InterviewQuestion;
-
-public interface InterviewQuestionRepository {
-    InterviewQuestion save(InterviewQuestion interviewQuestion);
-}
