--- src/main/java/KUSITMS/WITHUS/domain/application/interviewQuestion/service/InterviewQuestionService.java
@@ -1,7 +1,0 @@
-package KUSITMS.WITHUS.domain.application.interviewQuestion.service;
-
-import KUSITMS.WITHUS.domain.application.interviewQuestion.entity.InterviewQuestion;
-
-public interface InterviewQuestionService {
-    InterviewQuestion addQuestionToApplication(Long applicationId, Long userId, String content);
-}
