+++ src/main/java/KUSITMS/WITHUS/domain/interview/interviewQuestion/repository/InterviewQuestionJpaRepository.java
@@ -1,6 +1,6 @@
-package KUSITMS.WITHUS.domain.application.interviewQuestion.repository;
+package KUSITMS.WITHUS.domain.interview.interviewQuestion.repository;
 
-import KUSITMS.WITHUS.domain.application.interviewQuestion.entity.InterviewQuestion;
+import KUSITMS.WITHUS.domain.interview.interviewQuestion.entity.InterviewQuestion;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface InterviewQuestionJpaRepository extends JpaRepository<InterviewQuestion, Long> {
