+++ src/main/java/KUSITMS/WITHUS/domain/interview/interviewQuestion/repository/InterviewQuestionRepository.java
@@ -1,0 +1,7 @@
+package KUSITMS.WITHUS.domain.interview.interviewQuestion.repository;
+
+import KUSITMS.WITHUS.domain.interview.interviewQuestion.entity.InterviewQuestion;
+
+public interface InterviewQuestionRepository {
+    InterviewQuestion save(InterviewQuestion interviewQuestion);
+}
