+++ src/main/java/KUSITMS/WITHUS/domain/interview/interviewQuestion/service/InterviewQuestionService.java
@@ -1,0 +1,7 @@
+package KUSITMS.WITHUS.domain.interview.interviewQuestion.service;
+
+import KUSITMS.WITHUS.domain.interview.interviewQuestion.entity.InterviewQuestion;
+
+public interface InterviewQuestionService {
+    InterviewQuestion addQuestionToApplication(Long applicationId, Long userId, String content);
+}
